refactor MyFragment.kt into /util
