+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/PlaylistSongsFragment.kt
@@ -11,6 +11,7 @@ import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.util.MyFragment
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
